MMCPA-000: Search spellcheck warning.
If you you search for a term like `abcdef` that doesn't match any results and also doesn't return any suggestions, a warning is shown.

This module is hosted on Github and a PR was created there, https://github.com/nuvoleweb/search_api_spellcheck/pull/1.

The patch that matches that PR was included here in the patches folder and in composer.