MMCPA-000: Disable field check on Acquia Search.
search_api_solr_multilingual tries to see if the fields it's sending to solr exist before it sends them. This endpoint `schema/fieldtypes` is blocked on Acquia Search, so the system cannot index or retrieve content. For now, we can disable that check.

We should see why that is being blocked and if we can unblock that endpoint or the module should provide a way to configure if that should be checked.