AAVS-265: Move super category settings from override to install file.
Enable super category feature and menu block with config factory in install file.
As, Placing config in "override" folder didn't work. Because, we have an event
subscriber written for config save, which will always override the value
from the config placed in override folder and that will not allow drush
command, for super category feature, to execute properly.